make setup script modular

+++ setup.sh
@@ -1,31 +1,49 @@
 #!/bin/bash
 
-cd ~
-if [ -e .tmux.conf ]; then
-    rm -f .tmux.conf
-fi
-ln -s .dotfiles/tmuxconf .tmux.conf
+TARGET='__all__'
 
-if [ -e .gitconfig ]; then
-    rm -f .gitconfig
+if [ "$1" == '' ]; then
+    TARGET='__all__'
+elif [ "$1" == 'vim' ]; then
+    TARGET='vim'
+elif [ "$1" == 'bash' ]; then
+    TARGET='bash'
+else
+    echo "Invalid target"
+    exit 1
 fi
-ln -s .dotfiles/gitconfig .gitconfig
 
-if [ -e .bash_profile ]; then
-    rm -f .bash_profile
+if [ $TARGET == '__all__' -o $TARGET == 'bash' ]; then
+    cd ~
+    if [ -e .tmux.conf ]; then
+        rm -f .tmux.conf
+    fi
+    ln -s .dotfiles/tmuxconf .tmux.conf
+    
+    if [ -e .gitconfig ]; then
+        rm -f .gitconfig
+    fi
+    ln -s .dotfiles/gitconfig .gitconfig
+    
+    if [ -e .bash_profile ]; then
+        rm -f .bash_profile
+    fi
+    cp .dotfiles/bash_profile .bash_profile
 fi
-cp .dotfiles/bash_profile .bash_profile
 
-if [ -e .vimfiles ]; then
-    rm -rf .vimfiles
+if [ $TARGET == '__all__' -o $TARGET == 'vim' ]; then
+    cd ~
+    if [ -e .vimfiles ]; then
+        rm -rf .vimfiles
+    fi
+    git clone git@github.com:zwodahs/vimfiles .vimfiles
+    ln -s .vimfiles .vim
+    cd .vim
+    ./addlinks.sh
+    cd ~
+    
+    if [ -e .vimrc ]; then
+        rm -rf .vimrc
+    fi
+    ln -s .vim/vimrc .vimrc
 fi
-git clone git@github.com:zwodahs/vimfiles .vimfiles
-ln -s .vimfiles .vim
-cd .vim
-./addlinks.sh
-cd ~
-
-if [ -e .vimrc ]; then
-    rm -rf .vimrc
-fi
-ln -s .vim/vimrc .vimrc
